Refactor: split CompatibilityChangesTest into thematic test classes
Break the 3068-line monolithic CompatibilityChangesTest.java (108 tests)
into 10 focused, thematically organized test classes:

- ClassCompatibilityTest (11 tests): class-level changes
- SuperclassCompatibilityTest (12 tests): superclass hierarchy changes
- MethodCompatibilityTest (18 tests): method signature and modifier changes
- FieldCompatibilityTest (13 tests): field type, modifier and visibility changes
- ConstructorCompatibilityTest (6 tests): constructor changes
- InterfaceCompatibilityTest (19 tests): interface and default method changes
- ExceptionCompatibilityTest (5 tests): checked/unchecked exception changes
- AnnotationCompatibilityTest (15 tests): annotation changes on members
- GenericsCompatibilityTest (5 tests): generic type signature changes
- MemberRelocationCompatibilityTest (4 tests): members moved in hierarchy

All 108 tests preserved with identical logic; all pass.
